+++ Dto/DtoResolverTrait.php
@@ -15,9 +15,12 @@ namespace Wakeapp\Component\DtoResolver\Dto;
 
 use Symfony\Component\OptionsResolver\Exception\ExceptionInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+
 use function array_combine;
 use function array_intersect_key;
 use function array_keys;
+use function get_class;
+use function get_class_vars;
 use function get_object_vars;
 use function lcfirst;
 use function property_exists;
@@ -38,9 +41,6 @@ trait DtoResolverTrait
     private $definedProperties = [];
 
     /**
-     * @param array $data
-     * @param OptionsResolver|null $resolver
-     *
      * @throws ExceptionInterface
      */
     public function __construct(array $data = [], ?OptionsResolver $resolver = null)
@@ -71,18 +71,13 @@ trait DtoResolverTrait
         return $this->definedProperties;
     }
 
-    /**
-     * @return array
-     */
     public function jsonSerialize(): array
     {
         return $this->toArray();
     }
 
     /**
-     * @param bool $onlyDefinedData
-     *
-     * @return array
+     * TODO toArray with $onlyDefinedData = false returns only defined data, this bug need fix
      */
     public function toArray(bool $onlyDefinedData = true): array
     {
@@ -95,29 +90,29 @@ trait DtoResolverTrait
         return $data;
     }
 
-    /**
-     * @param OptionsResolver $resolver
-     */
     protected function configureOptions(OptionsResolver $resolver): void
     {
     }
 
-    /**
-     * @return array
-     */
     protected function getProperties(): array
     {
-        $data = $this->getObjectVars();
+        $data = $this->getClassVars();
 
         return array_keys($data);
     }
 
-    /**
-     * @return array
-     */
     protected function getObjectVars(): array
     {
         $data = get_object_vars($this);
+
+        unset($data['optionsResolver'], $data['definedProperties']);
+
+        return $data;
+    }
+
+    protected function getClassVars(): array
+    {
+        $data = get_class_vars(get_class($this));
 
         unset($data['optionsResolver'], $data['definedProperties']);
 
@@ -185,7 +180,6 @@ trait DtoResolverTrait
         $onlyDefinedData = $this->getOnlyDefinedData($normalizedData);
 
         $resolvedData = $this->getOptionResolver()->resolve($onlyDefinedData);
-
         foreach ($resolvedData as $propertyName => $value) {
             if (property_exists($this, $propertyName)) {
                 $this->$propertyName = $value;